+++ main.go
@@ -16,7 +16,7 @@ import (
 
 // Build information (set by ldflags)
 var (
-	version   = "5.5.3"
+	version   = "5.6.0"
 	buildTime = "unknown"
 	gitCommit = "unknown"
 )